requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-dateutil==2.8.2
pydantic==2.5.2
loguru==0.7.2
